refactor:BestClass,LatestClass,DeadlineClass 코드 겹치는 부분 공통 컴포넌트화
